requirements.txt
openapi
python-telegram-bot==13.5
lnd-grpc-client
python-dotenv
qrcode